refactor: RockScissorPaper, MukZziPpa 파일로 분리 및 게임이 정상 작동하도록 코드 변경
기존 main 파일의 두 게임을 RockScissorPaper, MukZziPpa 파일로 분리하였고 공통으로 사용하는 열거 타입을 Common으로 분리
